Added TASKS.txt
Added TASKS as a todo list for this project. Will attempt to seperate methods and implementations between the model, view and controller so that I know what goes where. I plan to work on the model portion first to get it all working. If I get bored with that I will work on the view concurrently and then do the controller last to tie the two together.