First big crack at the project!
I started filling out the classes. The basic parts were put into place for the GUI in scene builder. The minefield has not been constructed yet. This will mostly be done programmatically and I still need to decide how to do this and what to wrap it all in.
I started putting in method stubs to code up the Model of the game. Nothing is very complete yet but a lot of parts are put into place.