First code stub fill out
The View file has been fleshed out in Scene Builder with a basic 5x5 button grid and the needed buttons.
The Controller file has not really been touched yet. This will come later.
The Model file has been started but mostly has stub methods and a beginning to a constructor.