Completed View class as well as Minefield class
Minefield.java: Finished filling out methods for Minefield.java, may need to add more methods in the future but for now the current ones are working.

MineSweeperView: Removed the anchor pane from the center box on the the view. changed the initial grid pane to have 8x8 buttons instead of 5x5. this is more in convention with the original minesweeper. Medium and hard settings will come from this convention as well.

Minesweeper.java: was put back into play. Previously I was going to just use the controller to handle starting up Minefield but i think it will be cleaner if minesweeper.java handles that.

Controller: hasn’t really been touched yet. I will probably save this mostly for last.

Deleted a lot of the files that weren't supposed to be in the repos and will just make grading harder. The only files in here now should be the 4 or 5 that have code which will actually be graded.