Working on the controller
Started working on the controller. There are some things not working now. I deleted the
class Minesweeper.java since I decided to not use it and put the small
bit of code it was going to implement into the Controller.java.
The compiler is really complaining about this and i'm not sure if it will be
something I can tell it to reconfigure or if I will have to roll back to a previous build.